+++ gmsh2mixdv2/endian.h
@@ -1,8 +1,10 @@
 #ifndef ENDIAN_H
 #define ENDIAN_H
 
+#include <stdio.h>
+
 bool isBigEndian();
-void swapbytes(char *array, int nelem, int elsize);
+void swapbytes(char *array, size_t nelem, int elsize);
 void endianHandler(int invalue, char * buf);
